Rebase here

+++ Assets/Madduck/Scripts/Gameplay/GameData/WeightTable/Fish/FishWeightUtils.cs
@@ -46,7 +46,7 @@ namespace Madduck.GameData
         {
             return records.Select(record => record with
             {
-                Weight = record.Weight + _modifier(record)
+                Weight = (float)record.Weight + _modifier(record)
             }).ToList();
         }
     }
